[refactor](#45): muda os diretórios de /app para /src
