+++ src/app/(auth)/login/page.tsx
@@ -2,16 +2,16 @@
 
 import { LockIcon, UserIcon } from "lucide-react";
 import Link from "next/link";
-import { AutenticarUsuarioRequest } from "../../features/usuario/models/dto.model";
-import { LoginForm } from "../../features/usuario/models/form.model";
-import { authService } from "../../features/usuario/services/auth.service";
-import FyButton from "../../shared/components/fy-button/FyButton";
-import FyInput from "../../shared/components/fy-input/FyInput";
-import { useForm } from "../../shared/hooks/use-form";
+import { AutenticarUsuarioRequest } from "../../../features/usuario/models/dto.model";
+import { LoginForm } from "../../../features/usuario/models/form.model";
+import { authService } from "../../../features/usuario/services/auth.service";
+import FyButton from "../../../shared/components/fy-button/FyButton";
+import FyInput from "../../../shared/components/fy-input/FyInput";
+import { useForm } from "../../../shared/hooks/use-form";
 import { dependencies, validateLogin } from "./validation";
 import { useRouter } from "next/navigation";
-import { notificationService } from "../../shared/services/notification.service";
-import { getHttpErrorMessage } from "../../shared/utils/http-error";
+import { notificationService } from "../../../shared/services/notification.service";
+import { getHttpErrorMessage } from "../../../shared/utils/http-error";
 
 
 export default function Login() {
