+++ src/app/api/arquivos/route.ts
@@ -1,4 +1,4 @@
-import httpServer from "@/src/app/infrastructure/http/http-server";
+import httpServer from "@/src/infrastructure/http/http-server";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function GET(request: NextRequest) {
