+++ src/proxy.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { RefreshTokenRequest, TokenResponse } from "@/src/app/features/usuario/models/dto.model";
+import { RefreshTokenRequest, TokenResponse } from "@/src/features/usuario/models/dto.model";
 
 const publicOnlyRoutes = ["/login", "/register"];
 const API_URL = process.env.BACKEND_API_URL || "http://localhost:8080";
