+++ src/shared/components/fy-checkbox/FyCheckbox.spec.tsx
