+++ src/shared/components/fy-switch/FySwitch.spec.tsx
