+++ src/shared/utils/auth-client.ts
