+++ src/shared/utils/http-error.ts
